[dice] Refactor: Move DICE cert generation functions in a separate library
I am planning to use functions from
`//sw/device/silicon_creator/lib/cert:dice_mldsa` library to derive seed for
UDS MLDSA keygen in `sw/device/silicon_creator/manuf/base/ft_personalize.c`.
Linking `dice_mldsa` library with `ft_personalize.c` runs into failure since it
defines few functions from the already linked
`//sw/device/silicon_creator/lib/cert:dice` (like `dice_attest_cdi_0` and
`dice_attest_cdi_1`). Since `ft_personalize.c` only needs functions to build
and populate DICE certificate for UDS, CDI0, and CDI1, I am refactoring the
code to move these function definitions inside a new library target
`//sw/device/silicon_creator/lib/cert:dice_cert_build` so that they can be
linked along with `//sw/device/silicon_creator/lib/cert:dice_mldsa` when
building new `ft_personalize.c` with ML-DSA provisioning support

Tested by running
`//sw/host/provisioning/orchestrator/tests:e2e_emulation_cw340_test` and
`//sw/device/silicon_creator/rom_ext/e2e/attestation:print_certs_test_fpga_cw340_rom_ext`
locally
