+++ sw/device/silicon_creator/lib/cert/dice_cert_build.c
@@ -1,0 +1,202 @@
+#include "sw/device/lib/base/hardened.h"
+#include "sw/device/silicon_creator/lib/base/util.h"
+#include "sw/device/silicon_creator/lib/cert/cdi_0.h"  // Generated.
+#include "sw/device/silicon_creator/lib/cert/cdi_1.h"  // Generated.
+#include "sw/device/silicon_creator/lib/cert/cert.h"
+#include "sw/device/silicon_creator/lib/cert/dice.h"
+#include "sw/device/silicon_creator/lib/cert/template.h"
+#include "sw/device/silicon_creator/lib/cert/uds.h"  // Generated.
+#include "sw/device/silicon_creator/lib/drivers/hmac.h"
+#include "sw/device/silicon_creator/lib/drivers/lifecycle.h"
+#include "sw/device/silicon_creator/lib/error.h"
+#include "sw/device/silicon_creator/lib/otbn_boot_services.h"
+#include "sw/device/silicon_creator/lib/ownership/datatypes.h"
+#include "sw/device/silicon_creator/lib/sigverify/ecdsa_p256_key.h"
+
+const dice_cert_format_t kDiceCertFormat = kDiceCertFormatX509TcbInfo;
+
+static ecdsa_p256_signature_t curr_tbs_signature = {.r = {0}, .s = {0}};
+
+static uint8_t cdi_0_tbs_buffer[kCdi0MaxTbsSizeBytes];
+static uint8_t cdi_1_tbs_buffer[kCdi1MaxTbsSizeBytes];
+
+/**
+ * Returns true if debug (JTAG) access is exposed in the current LC state.
+ */
+static bool is_debug_exposed(void) {
+  lifecycle_state_t lc_state = lifecycle_state_get();
+  if (lc_state == kLcStateProd || lc_state == kLcStateProdEnd) {
+    return false;
+  }
+  return true;
+}
+
+/**
+ * Returns true if the OwnerSw is booting outside of prod ladder.
+ */
+static bool get_debug_mode_cdi1(owner_app_domain_t key_domain) {
+  if (launder32(key_domain) != kOwnerAppDomainProd) {
+    return true;
+  }
+  HARDENED_CHECK_EQ(key_domain, kOwnerAppDomainProd);
+  return false;
+}
+
+rom_error_t dice_uds_tbs_cert_build(
+    const hmac_digest_t *otp_creator_sw_cfg_measurement,
+    const hmac_digest_t *otp_owner_sw_cfg_measurement,
+    const hmac_digest_t *otp_rot_creator_auth_codesign_measurement,
+    const hmac_digest_t *otp_rot_creator_auth_state_measurement,
+    const cert_key_id_pair_t *key_ids,
+    const ecdsa_p256_public_key_t *uds_pubkey, uint8_t *tbs_cert,
+    size_t *tbs_cert_size) {
+  // Generate the TBS certificate.
+  uds_tbs_values_t uds_tbs_params = {0};
+
+  TEMPLATE_SET(uds_tbs_params, Uds, OtpCreatorSwCfgHash,
+               otp_creator_sw_cfg_measurement->digest);
+  TEMPLATE_SET(uds_tbs_params, Uds, OtpOwnerSwCfgHash,
+               otp_owner_sw_cfg_measurement->digest);
+  TEMPLATE_SET(uds_tbs_params, Uds, OtpRotCreatorAuthCodesignHash,
+               otp_rot_creator_auth_codesign_measurement->digest);
+  TEMPLATE_SET(uds_tbs_params, Uds, OtpRotCreatorAuthStateHash,
+               otp_rot_creator_auth_state_measurement->digest);
+  TEMPLATE_SET(uds_tbs_params, Uds, DebugFlag, is_debug_exposed());
+  TEMPLATE_SET(uds_tbs_params, Uds, CreatorPubKeyEcX, uds_pubkey->x);
+  TEMPLATE_SET(uds_tbs_params, Uds, CreatorPubKeyEcY, uds_pubkey->y);
+
+  TEMPLATE_SET_TRUNCATED(uds_tbs_params, Uds, CreatorPubKeyId,
+                         key_ids->cert->digest, kCertKeyIdSizeInBytes);
+  TEMPLATE_SET_TRUNCATED(uds_tbs_params, Uds, AuthKeyKeyId,
+                         key_ids->endorsement->digest, kCertKeyIdSizeInBytes);
+
+  HARDENED_RETURN_IF_ERROR(
+      uds_build_tbs(&uds_tbs_params, tbs_cert, tbs_cert_size));
+
+  return kErrorOk;
+}
+
+rom_error_t dice_cdi_0_cert_build(const hmac_digest_t *rom_ext_measurement,
+                                  uint32_t rom_ext_security_version,
+                                  const cert_key_id_pair_t *key_ids,
+                                  const ecdsa_p256_public_key_t *uds_pubkey,
+                                  const ecdsa_p256_public_key_t *cdi_0_pubkey,
+                                  uint8_t *cert, size_t *cert_size) {
+ 
+  hmac_digest_t rom_ext_hash = *rom_ext_measurement;
+  util_reverse_bytes(&rom_ext_hash, sizeof(rom_ext_hash));
+
+  uint32_t rom_ext_security_version_be =
+      __builtin_bswap32(rom_ext_security_version);
+
+  // Generate the TBS certificate.
+  cdi_0_tbs_values_t cdi_0_tbs_params = {0};
+
+  TEMPLATE_SET(cdi_0_tbs_params, Cdi0, RomExtHash, rom_ext_hash.digest);
+  TEMPLATE_SET(cdi_0_tbs_params, Cdi0, RomExtSecurityVersion,
+               &rom_ext_security_version_be);
+  TEMPLATE_SET(cdi_0_tbs_params, Cdi0, OwnerIntermediatePubKeyEcX,
+               cdi_0_pubkey->x);
+  TEMPLATE_SET(cdi_0_tbs_params, Cdi0, OwnerIntermediatePubKeyEcY,
+               cdi_0_pubkey->y);
+
+  TEMPLATE_SET_TRUNCATED(cdi_0_tbs_params, Cdi0, OwnerIntermediatePubKeyId,
+                         key_ids->cert->digest, kCertKeyIdSizeInBytes);
+  TEMPLATE_SET_TRUNCATED(cdi_0_tbs_params, Cdi0, CreatorPubKeyId,
+                         key_ids->endorsement->digest, kCertKeyIdSizeInBytes);
+
+  size_t tbs_size = kCdi0MaxTbsSizeBytes;
+  HARDENED_RETURN_IF_ERROR(
+      cdi_0_build_tbs(&cdi_0_tbs_params, cdi_0_tbs_buffer, &tbs_size));
+
+  // Sign the TBS and generate the certificate.
+  hmac_digest_t tbs_digest;
+  hmac_sha256(cdi_0_tbs_buffer, tbs_size, &tbs_digest);
+
+  ecdsa_p256_public_key_t uds_pubkey_le = *uds_pubkey;
+  util_reverse_bytes(uds_pubkey_le.x, sizeof(uds_pubkey_le.x));
+  util_reverse_bytes(uds_pubkey_le.y, sizeof(uds_pubkey_le.y));
+
+  HARDENED_RETURN_IF_ERROR(otbn_boot_attestation_endorse(
+      &tbs_digest, &curr_tbs_signature, &uds_pubkey_le));
+
+  util_p256_signature_le_to_be_convert(curr_tbs_signature.r,
+                                       curr_tbs_signature.s);
+
+  cdi_0_sig_values_t cdi_0_params = {0};
+  TEMPLATE_SET(cdi_0_params, Cdi0, Tbs, cdi_0_tbs_buffer);
+  // Asserts that tbs is fixed-length, so we don't need to set tbs_size.
+  TEMPLATE_ASSERT_FIXED_LENGTH(Cdi0, Tbs);
+  TEMPLATE_SET(cdi_0_params, Cdi0, CertSignatureR, curr_tbs_signature.r);
+  TEMPLATE_SET(cdi_0_params, Cdi0, CertSignatureS, curr_tbs_signature.s);
+
+  HARDENED_RETURN_IF_ERROR(cdi_0_build_cert(&cdi_0_params, cert, cert_size));
+
+  return kErrorOk;
+}
+
+rom_error_t dice_cdi_1_cert_build(
+    const hmac_digest_t *owner_measurement,
+    const hmac_digest_t *owner_manifest_measurement,
+    const hmac_digest_t *owner_history_hash, uint32_t owner_security_version,
+    owner_app_domain_t key_domain, const cert_key_id_pair_t *key_ids,
+    const ecdsa_p256_public_key_t *cdi_0_pubkey,
+    const ecdsa_p256_public_key_t *cdi_1_pubkey, uint8_t *cert,
+    size_t *cert_size) {
+  hmac_digest_t owner_hash = *owner_measurement;
+  hmac_digest_t owner_manifest_hash = *owner_manifest_measurement;
+  util_reverse_bytes(&owner_hash, sizeof(owner_hash));
+  util_reverse_bytes(&owner_manifest_hash, sizeof(owner_manifest_hash));
+
+  uint32_t owner_security_version_be =
+      __builtin_bswap32(owner_security_version);
+
+  // Generate the TBS certificate.
+  cdi_1_tbs_values_t cdi_1_tbs_params = {0};
+
+  TEMPLATE_SET(cdi_1_tbs_params, Cdi1, OwnerHash, owner_hash.digest);
+  TEMPLATE_SET(cdi_1_tbs_params, Cdi1, OwnerManifestHash,
+               owner_manifest_hash.digest);
+  TEMPLATE_SET(cdi_1_tbs_params, Cdi1, OwnerHistoryHash,
+               owner_history_hash->digest);
+  TEMPLATE_SET(cdi_1_tbs_params, Cdi1, OwnerSecurityVersion,
+               &owner_security_version_be);
+  TEMPLATE_SET(cdi_1_tbs_params, Cdi1, DebugFlag,
+               get_debug_mode_cdi1(key_domain));
+  TEMPLATE_SET(cdi_1_tbs_params, Cdi1, OwnerPubKeyEcX, cdi_1_pubkey->x);
+  TEMPLATE_SET(cdi_1_tbs_params, Cdi1, OwnerPubKeyEcY, cdi_1_pubkey->y);
+
+  TEMPLATE_SET_TRUNCATED(cdi_1_tbs_params, Cdi1, OwnerPubKeyId,
+                         key_ids->cert->digest, kCertKeyIdSizeInBytes);
+  TEMPLATE_SET_TRUNCATED(cdi_1_tbs_params, Cdi1, OwnerIntermediatePubKeyId,
+                         key_ids->endorsement->digest, kCertKeyIdSizeInBytes);
+
+  size_t tbs_size = kCdi1MaxTbsSizeBytes;
+  HARDENED_RETURN_IF_ERROR(
+      cdi_1_build_tbs(&cdi_1_tbs_params, cdi_1_tbs_buffer, &tbs_size));
+
+  // Sign the TBS and generate the certificate.
+  hmac_digest_t tbs_digest;
+  hmac_sha256(cdi_1_tbs_buffer, tbs_size, &tbs_digest);
+
+  ecdsa_p256_public_key_t cdi_0_pubkey_le = *cdi_0_pubkey;
+  util_reverse_bytes(cdi_0_pubkey_le.x, sizeof(cdi_0_pubkey_le.x));
+  util_reverse_bytes(cdi_0_pubkey_le.y, sizeof(cdi_0_pubkey_le.y));
+
+  HARDENED_RETURN_IF_ERROR(otbn_boot_attestation_endorse(
+      &tbs_digest, &curr_tbs_signature, &cdi_0_pubkey_le));
+
+  util_p256_signature_le_to_be_convert(curr_tbs_signature.r,
+                                       curr_tbs_signature.s);
+
+  cdi_1_sig_values_t cdi_1_params = {0};
+  TEMPLATE_SET(cdi_1_params, Cdi1, Tbs, cdi_1_tbs_buffer);
+  // Asserts that tbs is fixed-length, so we don't need to set tbs_size.
+  TEMPLATE_ASSERT_FIXED_LENGTH(Cdi1, Tbs);
+  TEMPLATE_SET(cdi_1_params, Cdi1, CertSignatureR, curr_tbs_signature.r);
+  TEMPLATE_SET(cdi_1_params, Cdi1, CertSignatureS, curr_tbs_signature.s);
+
+  HARDENED_RETURN_IF_ERROR(cdi_1_build_cert(&cdi_1_params, cert, cert_size));
+
+  return kErrorOk;
+}
